markup: Fix two bugs
The docs state that all chars marked with the accel
marker get an underline. But we were only underlining
the first in each text chunk.

Second, if an underline appears at the end of a text
chunk, or at the end of the text, we would just eat
it, which is unexpected.