fc: Drop a broken optimization
We were trying to cut short the wait for FcInit in
pango_fc_font_map_set_config. But there was a thinko:
while it is true that the fontmap which had set_config
called on it does not need to wait for FcInit anymore,
other threads fontmaps still do. Instead of making this
more complicated, just drop the optimization.