renderer: Fix averaging of strikethroughs
In 85bdfead1b36945db251 I made consecutive runs average
their strikeout positions, but I did handle the accounting
properly, causing previous runs to influence the strikeout
of later runs, if we have to draw them separately (due to
color change).

It would be nicer to average even across color changes and
draw the line continuously in that case as well, but that
would require two passes over the runs.

Fixes: #574