font-desc: Use local volatile double values to store sizes
Under i386 multiplying double and integer values and cast it to integer
may end up to compute the wrong value when only 387 FPU is used, because
the temporary value will be stored in a register whose precision isn't
good enough.
And so, some multiplications which are expected to produce an integer,
will actually return a truncated value, that will be eventually floored.

An example is 1.2 * 12800 that is clearly 15360.0, but will produce
15359 when casted to int in the said i386 environment.

So use a temporary double variable to ensure that we do this computation
in the double scope, before casting to int.

And this will avoid using the said register, even when using more
aggressive optimizations (as per marking the local variable volatile).

Fixes: https://gitlab.gnome.org/GNOME/pango/-/issues/580