break: Speed things up a bit
This is not revolutionary, but it helps a bit:
We don't need to check the line break rules one-by-one.
After we applied one, most of the following rules can
be skipped, since they would just do the same.

This is not ideal, since we still do far too many
checks when following the if-else cascade to the end,
but its a start.