Add line-height attributes
Add attributes for line-height, in a relative and
absolute variant.

This will be used to grow the logical extents of
runs in a way that is compatible with CSS semantics.

In markup, we support a new line_height attribute
that will be interpreted as absolute if it is an
integer > 1024, and as a relative factor otherwise.