Fix up logical rectangle reporting
Fix up the various apis that are reporting logical
rects to follow these rules:

Logical rectangles of lines are affected by the
line-height attribute, logical rectangles of smaller
units (runs, clusters, chars, cursors, etc) are not.

This fixes the text cursors and block cursors in
GTK to be 'normal' height, even when applying
line-height for double-spacing.