pangowin32-fontmap.c: Look harder for the matching font
When we look up our font from the list of fonts that we enumerated from
the system, follow what the CoreText backend does, so that if the font
gravity in the font description is getting in our way to find the font,
look for it again without the gravity.

This will enable us to find the font that really exists but was not found
due to FontDescription attributes.

Partially fixes #583.