pango/shape.c: Use the former rounding code as needed
The careful rounding code that was brought back as a result of commit
d4356779 only had effect if we are using a backend that supported font
hinting, so on other backends, use the former rounding code that was
dropped in favor of the careful rounding code that unfortunately did not
apply for all backends.

Should fix #578.