itemize: Implement emulated Small Caps
When we detect that Small Caps are requested, but not
available via OpenType font features, emulate Small Caps
by splitting the item into lowercase and uppercase runs
and add text transform and font scale attributes to the
lowercase runs to get the effect of Small Caps.

Still to do: resolve conflicts with preexisting text
transform attributes.