itemize: Emulate other casing variants
When we detect that one of the other casing variants
is requested, but not available via OpenType font
features, emulate it by splitting the item into
lowercase and uppercase runs and apply a suitable
font scale and text transform to the runs to get
the desired effect.

Still to do: resolve conflics with preexisting
text transform attributes.