Handle break-after-space correctly
When we are breaking after a space, we must not
count the width of the space towards the line,
since we are zeroing it later.

It is a bit annoying that there are multiple places
where this has to be taken into account.

Another missing bit in this code is that we are
only looking at a single whitespace character before
the break, when we should really look for a sequence
of spaces.