layout: Simplify things one more time
The key insight here is that if we find a broken item
does not fit, we are only interested in finding an
*earlier* breakpoint - the later ones aren't going
to produce a shorter run.

So we can just keep track of the last char we want
to allow breaking at.

This fixes the case of finding the minimum width
with wrap mode PANGO_WRAP_WORD.