Fix advance widths with transforms
The hb_font_t we were passing to Harfbuzz
for shaping was not taking context transforms
into account, and therefore was not scaling
advance widths as necessary. Always go through
pangos glyph extents to fix this.

Fixes: #620

This is a backport of commit ccb651dd2a876a4f4a4cb9351f05332173e709ba