Fix a thinko in zero_line_final_space
We must not add the width of the space back to
remaining_width, since we're just correcting the
glyphs to match the accounting that process_item
has already done.

This was showing up as justification operating
on wrong numbers when justifying lines with a
final space, leading to uneven margins.