Fix a case of unintended hyphenation
When we take a complete item because there are
no breakpoints, we were inserting a hyphen at
the end even though there may not be a breakpoint
there, and ending up with a hyphen in the middle
of an (overlong) line: Brat-wurst.

Test included.

This but was discovered with the help of
http://gitlab.gnome.org/matthiasc/layout-editor