serializer: Serialize non-ASCII correctly if char is unsigned
If char is unsigned, the first byte of a non-trivial UTF-8 sequence will
be 0x80 or higher, instead of being negative. In this case we need to
process it as UTF-8, instead of printing only the first byte and then
skipping to the next UTF-8 character.

char is usually signed on x86 and some other architectures, but not on
the ARM or PowerPC families, among others.

Resolves: https://gitlab.gnome.org/GNOME/pango/-/issues/652
Signed-off-by: Simon McVittie <smcv@debian.org>