wip: Add an api to select palettes for color glyphs
This commit adds a light-background attribute that can
be used to influence which palette will be used to render
COLRv0 layered glyphs with color fonts that have multiple
palettes. This needs a way for cairo to invalidate its
glyph cache when the selected palette changes, which
needs new freetype api.