Meson: Clean up HarfBuzz search
Pango 1.50.x and later require a HarfBuzz version that is buildable via
Meson, so we can just use the standard dependency() call for HarfBuzz
for all builds.

This means that we can drop all the items where we search for HarfBuzz
manually.

Also remove the stray HarfBuzz items that are no longer used.