docs: Don't install in doc/pango/reference folder
Currently Devhelp (and hence Builder) cannot find the pango documentation,
which means that all pango classes/enums links from GTK open the online
version of the pango docs instead of using the local one.
This is because pango installs its documentation in a subfolder under
$datadir/doc/pango/reference instead of $datadir/doc/pango*/ where
devhelp would be able to detect the pango*.devhelp2 file and hence
show the pango docs. I assume the reason to put those docs in a
subfolder is to keep the pango docs "tidy" since there can be 6 of them.
But not having them available in devhelp makes the whole point of having
docs installed moot, and as such I believe the inconvenient of having 6
different Pango* documentation in the devhelp list are less important
than being able to actually use the offline installed documentation.

Hence this commit changes the docs install directory to $datadir/doc.