pangowin32: register a "Sans-Serif" font
This creates a fake font family which will fall back to the first
available font in the alias list and as a result
pango_win32_font_map_load_font() will no longer fail for "sans-serif".

This was already done for other aliases, but for some reason not for
"sans-serif".

See #680