Drop a cursor height test
I can't find any justification why this should be
true nowadays, since we use the run height for the
cursor - it just isn't true the the runs in "one"
and "" have the same height.

Drop this test, and check in the other cursor height
test that empty lines still get non-zero cursors.