pango-font-private.h: Fix builds on Visual Studio+DWrite
The DirectWrite support code sadly needs to be done in C++, meaning that
more type checking is done by the Visual Studio compiler and one cannot do
some items in the C99 fashion (i.e. GCCism) as opposed to using plain C.

This fixes the build of pangodwrite-fontmap.cpp for the DWrite support.