pangocairo-font.c: Fix Windows DirectWrite build
Sadly, the APIs and COM types in dwrite.h that is shipped with the
Windows SDK at least is C++ only, so we can't just include it in C files
directly.  No version of the Windows SDK ship dwrite.h that is
compatible with C, unlike many of the DirectX headers that at least
shipped in the past with C support.

Instead, add a C++ source file that does the work to acquire the
cairo_font_face_t that we need from the DirectWrite IDWriteFontFace that
we added whe we populated the font map, and call that function from
pangocairo-font.c.  Make the definition of
create_cairo_font_face_for_hb_font() on DirectWrite builds marked with
static, as with the CoreText and FontConfig builds.

As a consequence, make sure pangocairo-private.h uses C linkage so that
things will link properly when DirectWrite is being used.