build: Fix HarfBuzz fallback for Windows
In order to build with DirectWrite support, HarfBuzz must also be built
with DirectWrite support, which is disabled by default.  Tell the
fallback build to look for DirectWrite and enabled it as a result-it
should be sufficiently present for Visual Studio with the Windows 8.0
SDK, and we are likely using a Windows 10 SDK since we require Visual
Studio 2015 or later for MSVC builds.

Remove the line in .gitlab-ci/test-msys2.sh that requests this support
since we are putting this in the main Meson build files.