README.win32: Update Windows build information
There is no more FontConfig support for Windows for Pango2, and let
people know they need DirectWrite support enabled in HarfBuzz and Cairo,
which may not be enabled by default.

Also let people know that if building HarfBuzz as a fallback as a part
of building Pango2, Visual Studio 2017 or later is required to
successfully build HarfBuzz.

Note that HarfBuzz DLLs built with Visual Studio 2017 or later should be
compatible with Pango built with Visual Studio 2015.