Redo the cairo backend code
We need to be more careful when creating a cairo
font face. Even on Windows or macOS, we may encounter
hb_font_t's that haven't been created from a native
font, so we need to handle those (if we have freetype).

Also simplify the cairo build machinery a bit, and
rely more on cairo-features.h.