pangowin32-utils.c: Add aliases as from Pango-1.x
This adds more font families for the PangoFontMap for the aliases that we are
using, as per what we did in Pango-1.x, sans the families that we share for
different Unicode needs, which would be used in fallback fontmaps.

Fallbacks would be added in a later commit, when it is clearer for how to
create a PangoFontMap for that purpose.