Support color palettes in fonts
Add a pango_context_set_palette to select whether
we prefer the default palette, the palette for light
background, or the palette for dark background. Also
add a palette attribute that can be used to override
this.

Make the cairo renderer pick up the palette, and
apply it when installing the font for a run.

Predefined palette names are "default", "light",
"dark", "palette0", "palette1", ...

Additionally, PangoHbFace can associate custom
names with palette indices.

To try this, use the new --palette option of
pango-view.

Fonts to try this with are Amiri Quran Colored
or the Bungee Color family.