PangoWin32: Query font descriptions using DirectWrite on LOGFONTW's
We now use DirectWrite to query the font weight, stretch, description
and so on by using first DirectWrite's GdiInterop to convert the
logfontw into an IDirectWriteFont, since the support in there for
querying font attributes are more complete in there.  Take out the
warned fonts items, as DirectWrite would support the font attributes
that we need much better (i.e. stretch and so on), and the warned fonts
items will get into the way.

Portions based on Luca Bacci's work for querying the font description
for Windows using DirectWrite for the upcoming Pango2.