meson.build: Check for GDI and DirectWrite support in HarfBuzz
These support are not enabled by default when building HarfBuzz, but if one or
both are enabled, we can simplify the codepath to create the needed hb_face_t.

This was not added previously as the needed HarfBuzz GDI API was not available
at the time the code was written and we are only integrating DirectWrite now.