Revert "PangoWin32: Use HarfBuzz platform APIs if available"
It turned out that there were issues in regards to building HarfBuzz as
a fallback dependency, so we retain using only the former method to
create the hb_face_t using only raw data, which does not depend on
platform API usage in HarfBuzz.

Hence, reduce the clutter in the build files a bit, which was a
necessary evil back in time.
This reverts commit 6b0aa77d23ac969c12eab00b178957a63befe5bd.