pangowin32-fontmap.c: Allow some support for loading custom fonts
Previously, when using GDI, we are able to add custom fonts to the system font
collection using AddFontResource()/AddFontResourceEx(), but this is not enough
as we transition to DirectWrite.

Allow some support for this for people that are using AddFontResource() (but
not yet AddFontResourceEx()), by using the GDI interop support in DirectWrite.

Let people know that currently AddFontResource() should be used for the moment
for this support to work, not AddFontResourceEx().