meson: Stop using fallback: kwarg for deps that don't need it
Subprojects that use meson.override_dependency() do not require the
caller to provide the subproject name and dependency variable name
inside the subproject.

All the dependencies we use can be switched in this way; including
gi-docgen which is now automatically executed by the find_program()
invocation thanks to `program_names` in the wrap file.