pangowin32: Create IDWriteFactory[3|5] objects if possible
If we have the dwrite_3.h header, first create the IDWriteFactory3 object,
which is available from Windows 10 or later, and grab the
IDWriteFactory5 object (available since Windows 10 Creators' Update) and
the original IDWriteFactory object from it, if we succeeded in doing so.

If we don't have dwrite_3.h (or IDWriteFactory3 from the system), create
the IDWriteFactory object as we did before.