pangowin32: Allow adding DWrite fontsets to fontmap
Extend the fontmap populate function a bit so that if we have added
custom DWrite font sets to the IDWriteFontSetBuilder[1] in our font map,
we could add them to the font collection that we obtained from
the system.

Note that this does not yet expose any public API for people to use to
add font files to the fontmap for Windows.