PangoWin32: Create hb_face_t using HarfBuzz platform APIs
...if the HarfBuzz build has the applicable Windows-specific APIs enabled. This
will simplify the codepath that is needed to create the hb_face_t for Pango
that we need.

Since the former method to create the hb_face_t in PangoWin32 appeared before
HarfBuzz added GDI support, and we only supported DirectWrite in Pango later,
only build the old code when neither DirectWrite nor GDI support in HarfBuzz
is found/enabled.