fonts.c: Round size / PANGO_SCALE to nearest int
The compiler/CRT may just well stuff the results of the division as a floating-
point number into the string for the font size that forms a part of the string
that we get from pango_font_description_to_string(), which is most probably not
what we want.

So, use round() on the results of that division if not using absolute font
sizes, so that we get the font size in a nicer looking manner when we call
pango_font_description_to_string().