build: Clean up Cairo dependency search and checks
Cairo 1.17.x is now built on Visual Studio using Meson, meaning that pkg-config
files are always around, plus 1.17.x contained many improvements, fixes and
features for Windows, so let's just limit looking for Cairo using pkg-config or
subprojects.

If we are building PangoFt2, also always check that Cairo contains FreeType and
FontConfig support (cairo-ft may or may not contain FontConfig support), or
if building Cairo as a subproject, build it with FreeType and FontConfig
enabled.

Consequently, this will simplify our checks on features that are enabled in
Cairo.