build: Use hb-directwrite more on Windows
As we are using more of DirectWrite in PangoWin32, it makes sense for us to use
hb-directwrite also when we are (unlikely) not enabling Cairo support for the
build, to reduce differences that may arise when we use GDI interop in this
process.

So, from this point, we use hb-gdi (or its raw form) only when we are enabling
Cairo support and do not have DirectWrite enabled in the Cairo build, or
HarfBuzz is built without DirectWrite support.