pangowin32-fontmap.c: Avoid double free in fontmap->finalize()
During the implementation of loading custom fonts, I missed the fact
that hash tables were free'ed in pango_win32_font_map_fini(), that was
used to reload fontmaps when a custom font is loaded.

This cause problems when we needed to create and destroy fontmaps
continually, as demonstrated in test-pangocairo-threads.