win32: Keep variations in PangoWin32Font
Keep a variations string in PangoWin32Font, and pass it
to both cairo when creating a cairo_scaled_font_t and to
harfbuzz when creating a hb_font_t. Unfortunately, the
dwrite code in cairo does not apply the variations atm.

We also restore the variations when describing the font.