pangowin32-dwrite-fontmap.cpp: Cleanup COM code a bit
We have pangowin32_release_com_obj() that was added to help implement
custom font loading for Windows 7/8.x, but this also makes releasing COM
objects here more or less like g_clear_object(), so let's use that more.

This is, sadly, for C++ for the moment, since DirectWrite code must be
in C++ as-is as the Windows SDK do not come with C-friendly headers for
DirectWrite, in any releases.