PangoWin32: Set up IDWriteTextAnalyzer[2] objects
We use the IDWriteTextAnalyzer2 object to find out what features are
supported by a IDWriteFontFace, which must be queried from a
IDWriteTextAnalyzer object that is created from the base IDWriteFactory
that we create.

Put these in the structure that we use to record the COM objects that
are related to the DWrite COM objects.