layout: Differentiate ellipsized and truncated
When the height of a multi-paragraph layout is limited, we may
end up not ellipsizing any line, but still omit entire paragraphs
due to height limitations. Introduce a new api,
pango_layout_is_truncated(), to determine if this is the case.
And take truncation into account when determining if we should
re-layout when the height changes.

Fixes: #823