Add PANGO_WRAP_NONE
This is an explicit way to disable wrapping.

Using the height limit of -1, which is the default, currently achieves
a similar effect (by limiting each paragraph to a single line of text,
which prevents wrapping), but we're planning to change the behavior
around limiting the number of lines to be per-layout instead of per-
paragraph.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>