layout: Fix measuring tab width
pango_shape_with_flags () wants item_text to be a substring of
paragraph_text in the address space sense; indeed, it has a pair of
g_return_if_fail's ensuring that this is the case. The C language
provides no such guarantee about string literals, the two identical
literals could point to the same string buffer or two disjoint string
buffers. Use an explicit variable to ensure the two pointers are
identical.

Fixes a critical in MSVC debug builds.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>