win32: Unify serial and dpi across fontmaps
Both the Win32CairoFontmap and Win32Fontmap track a serial and
dpi/resolution value. This causes conflicts in the cairo backends when
various functions take either of them.

So unify them by making the Cairo fontmap use the values from the win32
fontmap.

But unify the behavior of the dpi to conform to the Cairo fontmap
because that's the one that's actually in use and implements the
expected behavior.