win32: Turn font lookup failure warning into g_printerr()
Using a warning makes testsuites abort when they encounter an unknown
font.

However, the GTK testsuite tests proper fallback for unknown fonts in
various places and those cases abort on win32 currently.

We considered removing the warning entirely, but due to font lookup
differences between fontconfig and DWrite, being notified of a failure
is helpful.
That's why it's ultimately deemed to be the best solution for now to
still print the message to stderr, but not in a way that aborts the
program.