fc: Fix face lookup
The change in 75d63fb1c673de855 overlooked the fact that the style
name is often not present in the pattern. Instead of looking for
that, we need to walk the list of faces to find the one whose
weight and slant match the pattern of the font we're creating.

Fixes: #860